+++ packages/headless/src/app/reducers.ts
@@ -4,7 +4,6 @@ import {caseAssistConfigurationReducer} from '../features/case-assist-configurat
 import {caseContextReducer} from '../features/case-context/case-context-slice';
 import {caseFieldReducer} from '../features/case-field/case-field-slice';
 import {caseInputReducer} from '../features/case-input/case-input-slice';
-import {configurationReducer} from '../features/configuration/configuration-slice';
 import {contextReducer} from '../features/context/context-slice';
 import {debugReducer} from '../features/debug/debug-slice';
 import {versionReducer} from '../features/debug/version-slice';
@@ -36,7 +35,6 @@ import {queryReducer} from '../features/query/query-slice';
 import {questionAnsweringReducer} from '../features/question-answering/question-answering-slice';
 import {recentQueriesReducer} from '../features/recent-queries/recent-queries-slice';
 import {recentResultsReducer} from '../features/recent-results/recent-results-slice';
-import {recommendationReducer} from '../features/recommendation/recommendation-slice';
 import {resultPreviewReducer} from '../features/result-preview/result-preview-slice';
 import {searchHubReducer} from '../features/search-hub/search-hub-slice';
 import {searchReducer} from '../features/search/search-slice';
@@ -48,7 +46,6 @@ import {tabSetReducer} from '../features/tab-set/tab-set-slice';
 import {triggerReducer} from '../features/triggers/triggers-slice';
 import {undoable} from './undoable';
 
-export const configuration = configurationReducer;
 export const pagination = paginationReducer;
 export const facetSet = facetSetReducer;
 export const facetSearchSet = specificFacetSearchSetReducer;
@@ -93,7 +90,6 @@ export const history = undoable({
   },
   reducer: historyReducer,
 });
-export const recommendation = recommendationReducer;
 
 export const productListing = productListingReducer;
 export const sort = sortReducer;