+++ packages/headless/src/controllers/product-listing/results-per-page/headless-product-listing-results-per-page.test.ts
@@ -1,4 +1,5 @@
-import {configuration, pagination} from '../../../app/reducers';
+import {configuration} from '../../../app/common-reducers';
+import {pagination} from '../../../app/reducers';
 import {
   registerNumberOfResults,
   updateNumberOfResults,